services/docetl/src/requirements.txt
pandas
PyYAML
loguru
pdfminer.six
python-docx
openpyxl
